requirements.txt
twikit>=2.0.0
groq>=0.5.0
python-dotenv>=1.0.0
nest_asyncio>=1.6.0
schedule>=1.2.0
fake-useragent>=1.5.1
feedparser>=6.0.0
